@@ -28,7 +28,7 @@
     "postcss": "^8.4.32",
     "prettier": "^3.1.0",
     "prettier-plugin-svelte": "^3.2.6",
-    "svelte": "^5.0.0",
+    "svelte": "^5.17.4",
     "svelte-check": "^4.0.2",
     "svelte-virtual-scroll-list": "^1.3.0",
     "tailwindcss": "^3.3.6",
@@ -44,7 +44,8 @@
     "bigger-picture": "^1.1.17",
     "minisearch": "^7.0.0",
     "openseadragon": "^5.0.0",
-    "sander": "^0.6.0"
+    "sander": "^0.6.0",
+    "svelte-lightbox": "^1.1.6"
   },
   "description": "This website is a purely static build with a client-side search component. It is hosted by Github and was developed by [Sebastian Flick](https://github.com/flicksolutions) and [Peter Dängeli](https://github.com/pdaengeli) of the [Data Science Lab](https://github.com/dsl-unibe-ch) using Svelte and Svelte-Kit.",
   "main": "svelte.config.js",
